fix some stats 7f932138abfd4d57f8b38870cc741806378fcb46

+++ devtools/ymake/common/cyclestimer.cpp
@@ -16,7 +16,7 @@ size_t TCyclesTimer::GetUs() const {
 }
 
 double TCyclesTimer::GetSeconds() const {
-    return (::GetCycleCount() - Start_) / NHPTimer::GetCyclesPerSecond();
+    return double(::GetCycleCount() - Start_) / NHPTimer::GetCyclesPerSecond();
 }
 
 TCyclesTimerRestarter::TCyclesTimerRestarter(TCyclesTimer& cyclesTimer)
